Struct jobsteal::iter::Hide

pub struct Hide<T>(_);

Used to mask data so that implementations don't conflict.

Specifically, this is used in the implementation of SplitIterator for Zip, by hiding the base. This is because SplitIterator requires that the base be Split. However, for Split types, there is a blanket impl for SplitIterator for convenience, so if we didn't mask the type, there would be conflicting implementations of SplitIterator for Zip. This will be obsoleted when specialization becomes stable.

Numerous other iterator adapters also use this, but will all be cleared by specialization.