Trait jobsteal::iter::Consumer

pub trait Consumer<In: IntoIterator>: Sync {
    type Item;
    fn consume<C: Callback<Self::Item>>(&self, i: In, cb: C) -> C::Out;
}

A consumer takes an IntoIterator, which is usually the Base of a Spliterator, produces the desired iterator, and passes it to the callback given.

Associated Types

type Item

Required Methods

fn consume<C: Callback<Self::Item>>(&self, i: In, cb: C) -> C::Out

Consume the iterator, typically by passing it on to the parent consumer along with a callback which will receive a producer of items to transform.