Trait jobsteal::iter::ExactSizeSplitIterator

pub trait ExactSizeSplitIterator: SplitIterator {
    fn size(&self) -> usize;
}

An iterator for which the exact number of elements is known.

If this is implemented for an iterator,size_hint for that iterator must return a pair of the exact size.

Required Methods

fn size(&self) -> usize

Get the number of elements in this iterator.