Struct jobsteal::Spawner

pub struct Spawner<'pool, 'scope> {
    // some fields omitted
}

A job spawner associated with a specific scope. Jobs spawned using this may spawn new jobs with the same lifetime.

Methods

impl<'pool, 'scope> Spawner<'pool, 'scope>

fn submit<F>(&self, f: F) where F: 'scope + Send + FnOnce(&Spawner<'pool, 'scope>)

Execute a function which necessarily outlives the scope which this resides in.

fn scope<'new, F>(&self, f: F) where 'scope: 'new, F: FnOnce(&Spawner<'pool, 'new>)

Construct a new spawning scope smaller than the one this spawner resides in.

fn join<A, B, R_A, R_B>(&self, oper_a: A, oper_b: B) -> (R_A, R_B) where A: Send + FnOnce(&Spawner) -> R_A, B: Send + FnOnce(&Spawner) -> R_B, R_A: Send, R_B: Send

Execute two closures, possibly asynchronously, and return their results. This will block until they are both complete.