Struct jobsteal::WorkPool

pub struct WorkPool {
    // some fields omitted
}

The work pool manages worker threads in a work-stealing fork-join thread pool.

Methods

impl WorkPool

fn new(n: usize) -> Result<Self, Error>

Creates a work pool with n worker threads.

fn synchronize(&mut self)

Finish all current jobs which are queued, and synchronize the workers until it's time to start again.

If any of the threads have panicked, that panic will be propagated to this thread. Until the next job is submitted, workers will block, allowing other threads to have higher priority.

fn scope<'pool, 'new, F>(&'pool mut self, f: F) where F: FnOnce(&Spawner<'pool, 'new>)

Create a new spawning scope for submitting jobs. Any jobs submitted in this scope will be completed by the end of this function call.

fn submit<'a, F>(&'a mut self, f: F) where F: 'static + Send + FnOnce(&Spawner<'a, 'static>)

Execute a job which strictly owns its contents. The execution of this function may be deferred to beyond this function call. In the general case, it is safe to submit jobs which only strictly outlive the pool. However, there is always the chance that the pool could be leaked, violating the invariant that the job is completed before the borrow of the data is.

fn spin_up(&mut self)

Spin up all the workers, in case they were in a paused state.

fn spawner<'a>(&'a mut self) -> Spawner<'a, 'static>

Get a reference to the local spawner. This can only be used to spawn static jobs.